Implement menu
Use for now a blocking implementation for the button press detection.
This is not an optimal implementation, but the simplest one. Best would
be to move this code into a button "active" module, that generate long/short
push event. A shared, mutable event queue must be implemented for this first.